mcp-clients/stdio/python/python-basic-client/requirements.txt
mcp>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
